requirements.txt
paho-mqtt
pyyaml
rpi_ws281x
adafruit-circuitpython-neopixel
